+++ src/python/sinumber.rs
@@ -1,5 +1,6 @@
 use crate::{si::*, QuantityError};
 use numpy::{PyReadonlyArray1, PyReadonlyArray2, PyReadonlyArray3, PyReadonlyArray4};
+use pyo3::basic::CompareOp;
 use pyo3::exceptions::PyTypeError;
 use pyo3::prelude::*;
 use pyo3::PyNumberProtocol;
@@ -26,6 +27,17 @@ impl From<PySINumber> for SINumber {
 impl pyo3::class::basic::PyObjectProtocol for PySINumber {
     fn __repr__(&self) -> PyResult<String> {
         Ok(self.0.to_string())
+    }
+
+    fn __richcmp__(&self, other: Self, op: CompareOp) -> PyResult<bool> {
+        Ok(match op {
+            CompareOp::Eq => self.0 == other.0,
+            CompareOp::Ne => self.0 != other.0,
+            CompareOp::Lt => self.0 < other.0,
+            CompareOp::Le => self.0 <= other.0,
+            CompareOp::Gt => self.0 > other.0,
+            CompareOp::Ge => self.0 >= other.0,
+        })
     }
 }